practice/chapter8/code/requirements.txt
# BERTopic 토픽 모델링 필수 라이브러리
# 설치: pip install -r requirements.txt

# 핵심 라이브러리
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.33

# 데이터 처리
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 시각화
matplotlib>=3.5.0
plotly>=5.0.0

# 한국어 처리 (선택)
# konlpy>=0.6.0  # 형태소 분석기 (Mecab 별도 설치 필요)
